Setup rm to use the new parser.

+++ programs/rm.py
@@ -11,21 +11,23 @@ pa('-r', action="store_true", dest="recursive", default=False)
 pa('-v', action="store_true", dest="verbose", default=False)
 
 def run(shell, args):
+    parser.add_shell(shell)
     args = parser.parse_args(args)
-    if len(args.paths) >= 1:
-        for path in args.paths:
-            path = shell.iabs_path(path)
-            if fs.is_file(path) or (fs.is_directory(path) and args.recursive):
-                if args.verbose:
-                    shell.stdout.write("Removing %s" % (path, ))
-                fs.remove(path, recursive=args.recursive)
-            else:
-                if not args.recursive and fs.is_directory(path):
-                    shell.stderr.write("%s is a directory" % (path, ))
-                elif not fs.is_file(path):
-                    shell.stderr.write("%s is not a file" % (path,))
-    else:
-        shell.stderr.write("missing file operand")
+    if not parser.help:
+        if len(args.paths) >= 1:
+            for path in args.paths:
+                path = shell.iabs_path(path)
+                if fs.is_file(path) or (fs.is_directory(path) and args.recursive):
+                    if args.verbose:
+                        shell.stdout.write("Removing %s" % (path, ))
+                    fs.remove(path, recursive=args.recursive)
+                else:
+                    if not args.recursive and fs.is_directory(path):
+                        shell.stderr.write("%s is a directory" % (path, ))
+                    elif not fs.is_file(path):
+                        shell.stderr.write("%s is not a file" % (path,))
+        else:
+            shell.stderr.write("missing file operand")
 
 def help():
     return parser.format_help()
